refactor(view): move the service and tag lists out of view.go
The first column renders two interchangeable lists that share their selection
markers and state indicators, so they belong together and apart from the panels
they drive.

Declarations move unchanged.
